Added projects for entire domain

+++ Inventory.Repository.Base.Contracts/Models/CreditViewLineDt.cs
@@ -6,7 +6,7 @@ namespace Inventory.Repository.Base.Contracts.Models
     {
         public Guid CreditLineId { get; set; }
         public decimal CreditPrice { get; set; }
-        public decimal CreditQuantity { get; set; }
+        public int CreditQuantity { get; set; }
         public decimal TaxRate { get; set; }
     }
 }
